+++ platforms/defaults_unix_test.go
@@ -1,3 +1,6 @@
+//go:build !windows
+// +build !windows
+
 /*
    Copyright The containerd Authors.
